+++ client/src/Havoc/Demon/CommandOutput.cc
@@ -14,7 +14,7 @@ using namespace HavocNamespace::HavocSpace;
 
 void DispatchOutput::MessageOutput( QString JsonString, const QString& Date = "" ) const
 {
-    auto JsonDocument = QJsonDocument::fromJson( QByteArray::fromBase64( JsonString.toLocal8Bit( ) ) );
+    auto JsonDocument = QJsonDocument::fromJson( Util::base64_decode_capped( JsonString.toLocal8Bit( ) ) );
     auto TaskID       = JsonDocument[ "TaskID" ].toString();
     auto MessageType  = JsonDocument[ "Type" ].toString();
     auto Message      = JsonDocument[ "Message" ].toString();
@@ -57,7 +57,7 @@ void DispatchOutput::MessageOutput( QString JsonString, const QString& Date = ""
 
         if ( Type.compare( "screenshot" ) == 0 )
         {
-            auto DecodedData = QByteArray::fromBase64( Data.toLocal8Bit() );
+            auto DecodedData = Util::base64_decode_capped( Data.toLocal8Bit() );
             auto Name        = JsonDocument[ "MiscData2" ].toString();
 
             HavocX::Teamserver.TabSession->LootWidget->AddScreenshot( DemonCommandInstance->DemonID, Name, Date, DecodedData );
@@ -65,7 +65,7 @@ void DispatchOutput::MessageOutput( QString JsonString, const QString& Date = ""
         else if ( Type.compare( "download" ) == 0 )
         {
             auto MiscDataInfo = JsonDocument[ "MiscData2" ].toString().split( ";" );
-            auto Name         = QByteArray::fromBase64( MiscDataInfo[ 0 ].toLocal8Bit() );
+            auto Name         = Util::base64_decode_capped( MiscDataInfo[ 0 ].toLocal8Bit() );
             auto Size         = ( MiscDataInfo[ 1 ] );
 
             HavocX::Teamserver.TabSession->LootWidget->AddDownload( DemonCommandInstance->DemonID, Name, Size, Date, nullptr );
@@ -78,7 +78,7 @@ void DispatchOutput::MessageOutput( QString JsonString, const QString& Date = ""
                 {
                     if ( Session.ProcessList )
                     {
-                        auto Decoded = QByteArray::fromBase64( Data.toLocal8Bit() );
+                        auto Decoded = Util::base64_decode_capped( Data.toLocal8Bit() );
                         Session.ProcessList->UpdateProcessListJson( QJsonDocument::fromJson( Decoded ) );
                     }
                 }
@@ -92,7 +92,7 @@ void DispatchOutput::MessageOutput( QString JsonString, const QString& Date = ""
                 {
                     if ( Session.FileBrowser )
                     {
-                        auto Decoded = QByteArray::fromBase64( Data.toLocal8Bit() );
+                        auto Decoded = Util::base64_decode_capped( Data.toLocal8Bit() );
                         Session.FileBrowser->AddData( QJsonDocument::fromJson( Decoded ) );
                     }
                 }